move postgres into subfolder
