+++ drivers/entry.go
@@ -3,6 +3,7 @@ package drivers
 import (
 	"log"
 
+	"github.com/PumpkinSeed/sqlfuzz/drivers/postgres"
 	"github.com/PumpkinSeed/sqlfuzz/drivers/types"
 )
 
@@ -12,7 +13,7 @@ func New(f types.Flags) types.Driver {
 	case "mysql":
 		return MySQL{f: f}
 	case "postgres":
-		return Postgres{f: f}
+		return postgres.New(f)
 	default:
 		log.Fatal("Driver not implemented")
 		return nil
@@ -24,7 +25,7 @@ func NewTestable(f types.Flags) types.Testable {
 	case "mysql":
 		return MySQL{f: f}
 	case "postgres":
-		return Postgres{f: f}
+		return postgres.New(f)
 	default:
 		log.Fatal("Testable not implemented")
 		return nil
