+++ drivers/mysql.go
@@ -7,6 +7,7 @@ import (
 	"strings"
 
 	"github.com/PumpkinSeed/sqlfuzz/drivers/types"
+	"github.com/PumpkinSeed/sqlfuzz/drivers/utils"
 )
 
 const (
@@ -17,7 +18,7 @@ const (
 var (
 	mySQLNameToTestCase = map[string]types.TestCase{
 		"single": {
-			TableToCreateQueryMap: map[string]string{DefaultTableCreateQueryKey: `CREATE TABLE %s (
+			TableToCreateQueryMap: map[string]string{utils.DefaultTableCreateQueryKey: `CREATE TABLE %s (
 		id INT(6) UNSIGNED,
 		firstname VARCHAR(30),
 		lastname VARCHAR(30),
@@ -209,7 +210,7 @@ func (m MySQL) MultiDescribe(tables []string, db *sql.DB) (map[string][]types.Fi
 	processedTables := make(map[string]struct{})
 	tableToDescriptorMap := make(map[string][]types.FieldDescriptor)
 	for {
-		newTableToDescriptorMap, newlyReferencedTables, err := multiDescribeHelper(tables, processedTables, db, m)
+		newTableToDescriptorMap, newlyReferencedTables, err := utils.MultiDescribeHelper(tables, processedTables, db, m)
 		if err != nil {
 			return nil, nil, err
 		}
@@ -221,7 +222,7 @@ func (m MySQL) MultiDescribe(tables []string, db *sql.DB) (map[string][]types.Fi
 		}
 		tables = newlyReferencedTables
 	}
-	insertionOrder, err := getInsertionOrder(tableToDescriptorMap)
+	insertionOrder, err := utils.GetInsertionOrder(tableToDescriptorMap)
 	if err != nil {
 		return nil, nil, err
 	}
@@ -245,7 +246,7 @@ func (MySQL) GetLatestColumnValue(table, column string, db *sql.DB) (interface{}
 
 // TestTable only for test purposes
 func (m MySQL) TestTable(db *sql.DB, testCase, table string) error {
-	return testTable(db, testCase, table, m)
+	return utils.TestTable(db, testCase, table, m)
 }
 
 func (MySQL) GetTestCase(name string) (types.TestCase, error) {
