+++ drivers/postgres/postgres_example_test.go
@@ -1,4 +1,4 @@
-package drivers
+package postgres
 
 import (
 	"database/sql"
